+++ 4scanner.egg-info/requires.txt
@@ -1,0 +1,1 @@
+requests
