+++ bin/4scanner
@@ -3,7 +3,6 @@
 import argparse
 import os
 import scanner
-from scanner import scanner
 import time
