+++ hash_test_file.txt
@@ -1,0 +1,1 @@
+some_hash
