+++ hash_test_file2.txt
@@ -1,0 +1,2 @@
+
+not checking for duplicate
