+++ test/test.py
@@ -1,26 +1,25 @@
-#!/usr/bin/env python3
-
 import json
 import os
-from scanner import scanner, dupecheck, chan_info, download
+import six
+from scanner import scann, chan_info , download ,dupecheck
 
 print("Testing scanner.py")
 
 print("--------------------------------------------------------")
 print("Testing: scanner.get_catalog_json                      -")
 print("--------------------------------------------------------")
-catalog_json = scanner.get_catalog_json("a", "4chan")
-print('\x1b[6;30;42m' + 'OK' + '\x1b[0m')
+catalog_json = scann.get_catalog_json("a", "4chan")
+six.print_(('\x1b[6;30;42m' + 'OK' + '\x1b[0m'))
 
 print("--------------------------------------------------------")
 print("Testing: scanner.scan_thread                           -")
 print("--------------------------------------------------------")
 
-list_of_threads = scanner.scan_thread("anime", catalog_json)
+list_of_threads = scann.scan_thread("anime", catalog_json)
 for i in list_of_threads:
      print(i)
 
-print('\x1b[6;30;42m' + 'OK' + '\x1b[0m')
+six.print_(('\x1b[6;30;42m' + 'OK' + '\x1b[0m'))
 
 print("--------------------------------------------------------")
 print("!!! scanner.download_thread not tested yet !!!         -")
@@ -30,9 +29,9 @@ print("--------------------------------------------------------")
 print("Testing: scanner.add_to_downloaded                     -")
 print("--------------------------------------------------------")
 
-scanner.add_to_downloaded("139294614", "log", ".")
+scann.add_to_downloaded("139294614", "log", ".")
 
-print('\x1b[6;30;42m' + 'OK' + '\x1b[0m')
+six.print_(('\x1b[6;30;42m' + 'OK' + '\x1b[0m'))
 
 print("--------------------------------------------------------")
 print("Testing: scanner.folder_size_mb                        -")
@@ -44,12 +43,12 @@ os.system("mkdir folder_size_test/subfolder")
 os.system("dd if=/dev/zero of=folder_size_test/subfolder/15mbfile bs=15000000 count=1")
 
 # Getting folder size
-size = scanner.folder_size_mb("folder_size_test")
+size = scann.folder_size_mb("folder_size_test")
 if int(size) != 15:
     print(size)
     exit(1)
 
-print('\x1b[6;30;42m' + 'OK' + '\x1b[0m')
+six.print_(('\x1b[6;30;42m' + 'OK' + '\x1b[0m'))
 
 print("--------------------------------------------------------")
 print("!!! scanner.check_quota not tested yet !!!             -")
@@ -57,10 +56,10 @@ print("--------------------------------------------------------")
 
 print("--------------------------------------------------------")
 print("Testing:                                               -")
-print("scanner.get_check_duplicate                            -")
-print("scanner.get_condition                                  -")
-print("scanner.get_imageboard                                 -")
-print("scanner.get_keyword                                    -")
+print("scann.get_check_duplicate                            -")
+print("scann.get_condition                                  -")
+print("scann.get_imageboard                                 -")
+print("scann.get_keyword                                    -")
 print("--------------------------------------------------------")
 
 json_file = json.load(open("test/test_config.json"))
@@ -71,10 +70,10 @@ for search in json_file["searches"]:
     search_list.append(search)
 
 # Test on the first search with all optionals parameters
-duplicate1  = scanner.get_check_duplicate(search_list[0])
-condition1  = scanner.get_condition(search_list[0])
-imageboard1 = scanner.get_imageboard(search_list[0])
-keyword1    = scanner.get_keyword(search_list[0])
+duplicate1  = scann.get_check_duplicate(search_list[0])
+condition1  = scann.get_condition(search_list[0])
+imageboard1 = scann.get_imageboard(search_list[0])
+keyword1    = scann.get_keyword(search_list[0])
 
 if not duplicate1:
     print("duplicate1 should be True but is False")
@@ -104,10 +103,10 @@ if keyword1 != ['keyword1', 'keyword2', 'keyword3']:
     print("keyword1 should be equal to ['keyword1', 'keyword2', 'keyword3']")
     exit(1)
 
-duplicate2  = scanner.get_check_duplicate(search_list[1])
-condition2  = scanner.get_condition(search_list[1])
-imageboard2 = scanner.get_imageboard(search_list[1])
-keyword2    = scanner.get_keyword(search_list[1])
+duplicate2  = scann.get_check_duplicate(search_list[1])
+condition2  = scann.get_condition(search_list[1])
+imageboard2 = scann.get_imageboard(search_list[1])
+keyword2    = scann.get_keyword(search_list[1])
 
 if not duplicate2:
     print("duplicate2 should be True but is False")
@@ -137,13 +136,13 @@ if keyword2 != ['keyword']:
     print("keyword2 should be equal to ['keyword']")
     exit(1)
 
-print('\x1b[6;30;42m' + 'OK' + '\x1b[0m')
+six.print_(('\x1b[6;30;42m' + 'OK' + '\x1b[0m'))
 
 print("--------------------------------------------------------")
 print("!!! scanner.scan not tested yet !!!                    -")
 print("--------------------------------------------------------")
 
-print('\x1b[6;30;42m' + 'All test OK for scanner.py' + '\x1b[0m')
+six.print_(('\x1b[6;30;42m' + 'All test OK for scanner.py' + '\x1b[0m'))
 
 print("Testing dupecheck.py")
 
@@ -156,7 +155,7 @@ if hash != "b3ce9cb3aefc5e240b4295b406ce8b9a":
     print("hash should be b3ce9cb3aefc5e240b4295b406ce8b9a")
     exit(1)
 
-print('\x1b[6;30;42m' + 'OK' + '\x1b[0m')
+six.print_(('\x1b[6;30;42m' + 'OK' + '\x1b[0m'))
 
 print("--------------------------------------------------------")
 print("Testing: dupecheck.add_to_file                         -")
@@ -170,7 +169,7 @@ if 'some_hash_to_write' not in open("hash_test_file.txt").read():
     print("the string was not written to hash_test_file.txt")
     exit(1)
 
-print('\x1b[6;30;42m' + 'OK' + '\x1b[0m')
+six.print_(('\x1b[6;30;42m' + 'OK' + '\x1b[0m'))
 
 print("--------------------------------------------------------")
 print("Testing: dupecheck.is_duplicate                        -")
@@ -190,7 +189,7 @@ if is_dupe:
     print("is_dupe should be False")
     exit(1)
 
-print('\x1b[6;30;42m' + 'OK' + '\x1b[0m')
+six.print_(('\x1b[6;30;42m' + 'OK' + '\x1b[0m'))
 
 print("--------------------------------------------------------")
 print("Testing: dupecheck.hash_img_in_folder                  -")
@@ -220,9 +219,9 @@ if 'not checking for duplicate' not in open("hash_test_file2.txt").read():
     print("'not checking for duplicate' should be in hash_test_file2.txt")
     exit(1)
 
-print('\x1b[6;30;42m' + 'OK' + '\x1b[0m')
+six.print_(('\x1b[6;30;42m' + 'OK' + '\x1b[0m'))
 
-print('\x1b[6;30;42m' + 'All test OK for dupecheck.py' + '\x1b[0m')
+six.print_(('\x1b[6;30;42m' + 'All test OK for dupecheck.py' + '\x1b[0m'))
 
 print("Testing chan_info.py")
 
@@ -282,9 +281,9 @@ if info_uboachan[3] != "https://uboachan.net/":
     print("chan_image_base_url wrong for uboachan")
     exit(1)
 
-print('\x1b[6;30;42m' + 'OK' + '\x1b[0m')
+six.print_(('\x1b[6;30;42m' + 'OK' + '\x1b[0m'))
 
-print('\x1b[6;30;42m' + 'All test OK for chan_info.py' + '\x1b[0m')
+six.print_(('\x1b[6;30;42m' + 'All test OK for chan_info.py' + '\x1b[0m'))
 
 print("Testing download.py")
 
@@ -305,7 +304,7 @@ if not os.path.exists("test_create_dir"):
 # Testing again because the function should not crash if folder already exist
 download.create_dir("test_create_dir")
 
-print('\x1b[6;30;42m' + 'OK' + '\x1b[0m')
+six.print_(('\x1b[6;30;42m' + 'OK' + '\x1b[0m'))
 
 print("--------------------------------------------------------")
 print("Testing: download.add_to_downloaded_log                -")
@@ -328,7 +327,7 @@ if downloaded:
     print("'returned' should be False")
     exit(1)
 
-print('\x1b[6;30;42m' + 'OK' + '\x1b[0m')
+six.print_(('\x1b[6;30;42m' + 'OK' + '\x1b[0m'))
 
 print("--------------------------------------------------------")
 print("Testing: download.extension_condition                  -")
@@ -342,7 +341,7 @@ if download.extension_condition([".jpg"], ".png"):
     print("different extension should return False")
     exit(1)
 
-print('\x1b[6;30;42m' + 'OK' + '\x1b[0m')
+six.print_(('\x1b[6;30;42m' + 'OK' + '\x1b[0m'))
 
 print("--------------------------------------------------------")
 print("Testing: download.filename_condition                   -")
@@ -356,7 +355,7 @@ if download.filename_condition(["PIC"], "IMG_2345.jpg"):
     print("PIC is not in IMG_2345, should return False")
     exit(1)
 
-print('\x1b[6;30;42m' + 'OK' + '\x1b[0m')
+six.print_(('\x1b[6;30;42m' + 'OK' + '\x1b[0m'))
 
 print("--------------------------------------------------------")
 print("Testing: download.width_condition                      -")
@@ -386,7 +385,7 @@ if download.width_condition("<100", 101):
     print("101 is not lower than 100, should be False")
     exit(1)
 
-print('\x1b[6;30;42m' + 'OK' + '\x1b[0m')
+six.print_(('\x1b[6;30;42m' + 'OK' + '\x1b[0m'))
 
 
 print("--------------------------------------------------------")
@@ -409,7 +408,7 @@ if download.all_condition_check(all_false):
     print("all conditions are False, should return False")
     exit(1)
 
-print('\x1b[6;30;42m' + 'OK' + '\x1b[0m')
+six.print_(('\x1b[6;30;42m' + 'OK' + '\x1b[0m'))
 
 print("--------------------------------------------------------")
 print("!!! download.meet_dl_condition not tested yet !!!      -")
@@ -436,7 +435,7 @@ if not os.path.isfile("not_duplicate.png"):
     print("not_duplicate.png should not have been deleted.")
     exit(1)
 
-print('\x1b[6;30;42m' + 'OK' + '\x1b[0m')
+six.print_(('\x1b[6;30;42m' + 'OK' + '\x1b[0m'))
 
 print("--------------------------------------------------------")
 print("Testing: download.download_image                       -")
@@ -450,11 +449,11 @@ if not os.path.isfile(file_path):
     print("4scanner128.png should have been downloaded.")
     exit(1)
 
-print('\x1b[6;30;42m' + 'OK' + '\x1b[0m')
+six.print_(('\x1b[6;30;42m' + 'OK' + '\x1b[0m'))
 
 print("--------------------------------------------------------")
 print("!!! download.download_thread not tested yet !!!        -")
 print("--------------------------------------------------------")
 
-print('\x1b[6;30;42m' + 'All test OK for download.py' + '\x1b[0m')
-print('\x1b[6;30;42m' + 'SUCCESS' + '\x1b[0m')
+six.print_(('\x1b[6;30;42m' + 'All test OK for download.py' + '\x1b[0m'))
+six.print_(('\x1b[6;30;42m' + 'SUCCESS' + '\x1b[0m'))
