+++ test_hash_log_not_dupe.txt
@@ -1,0 +1,2 @@
+not_our_img_md5
+b3ce9cb3aefc5e240b4295b406ce8b9a
